committing shitty code because I need to pack my desktop for Toronto
